+++ inc/options/Controller.php
@@ -26,8 +26,7 @@ class Controller
     // new \SFX\Options\AdminOptionPages();
     new \SFX\Options\ACF\OptionsContact();
     new \SFX\Options\ACF\OptionsSocialMedia();
-    // new \SFX\Options\ACF\OptionsGeneral();
-    new \SFX\Options\ACF\OptionsLogo();
+  
     new \SFX\Options\ACF\OptionsCustomScripts();
     new \SFX\Options\ACF\OptionsPresetScripts();
     new \SFX\Options\ACF\OptionsHeader();